==> Codes Provided/4. Tamer _POS Analysis.Rmd <==
---
title: "POS Analysis"
author: "Walid Gomaa, Mohamed A. Khamis"
date: "17/09/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## POS Analysis

```{r}
library(reticulate)

use_condaenv(conda_list()$name, required = TRUE)
```

```{r}
songs <- read.csv("./data/songs.csv", header = TRUE, sep = ",", 
                  stringsAsFactors = FALSE, encoding = "UTF-8", na.strings = "") # Empty fields are replaced by NA.
```

```{python}
from camel_tools.disambig.mle import MLEDisambiguator
from camel_tools.tagger.default import DefaultTagger

py_songs = r.songs
```

```{python}
mled = MLEDisambiguator.pretrained()
tagger = DefaultTagger(mled, 'pos')
```

```{python}
py_songs["POS Tags"] = [tagger.tag(ly.split()) for ly in py_songs["Lyrics"]]
```

```{python}
py_songs["POS Tags"]
```

```{python}
import pandas as pd

tags = ['noun', 'noun_prop', 'verb', 'prep', 'adj', 'part_voc', 'pron', 'noun_quant', 'conj']

#py_songs["test"] = [pd.Series(p).value_counts() for p in py_songs["POS Tags"]] 
py_songs["Feature Vector"] = [{t:p.count(t) for t in set(tags)} for p in py_songs["POS Tags"]]
```

```{python}
py_songs["Feature Vector"][0]
py_songs["Feature Vector"][10]
py_songs["Feature Vector"][100]
#py_songs["test"][0].to_dict()
```

In the following cell we extract the feature vectors based on POS features and put them into a 2-dim array.

```{python}
import numpy as np

N = len(py_songs)
M = len(tags)

feature_matrix = np.zeros((N,M))

for i in range(N):
  feature_matrix[i,:] = list(py_songs["Feature Vector"][i].values())

#ss[1,:] =list(py_songs["Feature Vector"][100].values())
#ss = np.array(list(py_songs["Feature Vector"][100].values()))
#ss
#ss.shape

feature_matrix
```

**Standardization**: Standardization scales, or shifts, the values for each numerical feature in your dataset so that the features have a mean of 0 and standard deviation of 1:

```{python}
from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
scaled_feature_matrix = scaler.fit_transform(feature_matrix)

scaled_feature_matrix
```

Applying the kmeans cluster algorithm.

```{python}
no_clust = 3

kmeans = KMeans(init = "random",
                n_clusters = no_clust,
                n_init = 10,
                max_iter=300,
                random_state=42)
```

```{python}
kmeans.fit(scaled_feature_matrix)
```

```{python}
idx = [i for i in range(N) if kmeans.labels_[i] == 2]
idx
```

```{python}
ss = [py_songs["Song"][i] for i in range(N) if i in idx]
ss
```

# ========================================================================================================================================================================================================================================================================

========================================================================================================================================================================================================================================================================
